Moves fsm_runner stuff into runner.{h,c}
