+++ src/retest/Makefile
@@ -3,13 +3,14 @@
 SRC += src/retest/main.c
 SRC += src/retest/reperf.c
 SRC += src/retest/cvtpcre.c
+SRC += src/retest/runner.c
 
 .for src in ${SRC:Msrc/retest/main.c} ${SRC:Msrc/retest/reperf.c} ${SRC:Msrc/retest/cvtpcre.c}
 CFLAGS.${src} += -I src # XXX: for internal.h
 DFLAGS.${src} += -I src # XXX: for internal.h
 .endfor
 
-.for src in ${SRC:Msrc/retest/reperf.c} ${SRC:Msrc/retest/cvtpcre.c}
+.for src in ${SRC:Msrc/retest/main.c} ${SRC:Msrc/retest/reperf.c} ${SRC:Msrc/retest/cvtpcre.c} ${SRC:Msrc/retest/runner.c}
 CFLAGS.${src} += -std=c99
 DFLAGS.${src} += -std=c99
 .endfor
@@ -21,6 +22,12 @@ LFLAGS.reperf += -ldl
 PROG += retest
 PROG += reperf
 PROG += cvtpcre
+
+.for prg in ${PROG}
+${BUILD}/bin/${prg}: ${BUILD}/src/retest/runner.o
+.endfor
+
+${BUILD}/lib/runner.o: $
 
 .for lib in ${LIB:Mlibfsm} ${LIB:Mlibre}
 ${BUILD}/bin/retest: ${BUILD}/lib/${lib:R}.a
