+++ src/retest/reperf.c
@@ -10,7 +10,6 @@
 #endif
 
 #include <unistd.h>
-#include <dlfcn.h>
 
 #include <assert.h>
 #include <errno.h>
@@ -37,6 +36,7 @@
 #include "libre/class.h" /* XXX */
 #include "libre/ast.h" /* XXX */
 
+#include "runner.h"
 
 #define DEBUG_ESCAPES     0
 #define DEBUG_VM_FSM      0
@@ -85,30 +85,9 @@ struct str {
 	size_t cap;
 };
 
-enum error_type {
-	ERROR_NONE = 0          ,
-	/* ERROR_BAD_RECORD_TYPE   , */
-	/* ERROR_ESCAPE_SEQUENCE   , */
-	ERROR_PARSING_REGEXP    ,
-	ERROR_DETERMINISING     ,
-	ERROR_COMPILING_BYTECODE,
-	ERROR_SHOULD_MATCH      ,
-	ERROR_SHOULD_NOT_MATCH  ,
-	ERROR_FILE_IO           ,
-	ERROR_CLOCK_ERROR       ,
-	ERROR_INVALID_PARAMETER ,
-};
-
 enum match_type {
 	MATCH_STRING = 0,
 	MATCH_FILE   = 1
-};
-
-enum implementation {
-	IMPL_C,
-	IMPL_VMC,
-	IMPL_VMASM,
-	IMPL_INTERPRET
 };
 
 struct perf_case {
@@ -507,261 +486,6 @@ read_file(const char *path, struct str *contents)
 	return 0;
 }
 
-struct fsm_runner {
-	enum implementation impl;
-
-	union {
-		struct {
-			void *h;
-			int (*func)(const char *, const char *);
-		} impl_c;
-
-		struct {
-			void *h;
-			int (*func)(const unsigned char *, size_t);
-		} impl_asm;
-
-		struct {
-			struct fsm_dfavm *vm;
-		} impl_vm;
-	} u;
-};
-
-static enum error_type
-runner_init_compiled(struct fsm *fsm, struct fsm_runner *r, enum implementation impl)
-{
-	static fsm_print *asm_print = fsm_print_vmasm_amd64_att;
-
-	char tmp_c[]  = "/tmp/fsmcompile_c-XXXXXX";
-	char tmp_o[]  = "/tmp/fsmcompile_o-XXXXXX";
-	char tmp_so[] = "/tmp/fsmcompile_so-XXXXXX";
-
-	char cmd[sizeof tmp_c + sizeof tmp_o + sizeof tmp_so + 256];
-	const char *cc, *cflags, *as, *asflags;
-	int fd_c, fd_so, fd_o;
-	FILE *f = NULL;
-	void *h = NULL;
-
-	/*
-	 * The IO API would depend on c->mt == MATCH_STRING ? FSM_IO_PAIR : FSM_IO_GETC;
-	 * except we read in the entire file below anyway, so we just
-	 * use FSM_IO_PAIR for both. The type of fsm_main depends on the IO API.
-	 */
-	opt.io = FSM_IO_PAIR;
-
-	fd_c  = mkstemp(tmp_c);
-	fd_so = mkstemp(tmp_so);
-	fd_o  = -1;
-
-	f = fdopen(fd_c, "w");
-	if (f == NULL) {
-		perror(tmp_c);
-		return ERROR_FILE_IO;
-	}
-
-	switch (impl) {
-	case IMPL_C:     fsm_print_c(f, fsm);   break;
-	case IMPL_VMC:   fsm_print_vmc(f, fsm); break;
-	case IMPL_VMASM:
-			 {
-				 asm_print(f,fsm);
-			 }
-			 break;
-
-	case IMPL_INTERPRET:
-			 assert(!"should not reach!");
-			 break;
-	}
-
-	if (EOF == fflush(f)) {
-		perror(tmp_c);
-		return ERROR_FILE_IO;
-	}
-
-	cc     = getenv("CC");
-	cflags = getenv("CFLAGS");
-
-	switch (impl) {
-	case IMPL_C:
-	case IMPL_VMC:
-		(void) snprintf(cmd, sizeof cmd, "%s %s -xc -shared -fPIC %s -o %s",
-				cc ? cc : "gcc", cflags ? cflags : "-std=c89 -pedantic -Wall -O3",
-				tmp_c, tmp_so);
-
-		if (0 != system(cmd)) {
-			perror(cmd);
-			return ERROR_FILE_IO;
-		}
-
-		break;
-
-	case IMPL_VMASM:
-		as      = getenv("AS");
-		asflags = getenv("ASFLAGS");
-
-		fd_o = mkstemp(tmp_o);
-
-		(void) snprintf(cmd, sizeof cmd, "%s %s -o %s %s",
-				as ? as : "as", asflags ? asflags : "", tmp_c, tmp_o);
-
-		if (0 != system(cmd)) {
-			perror(cmd);
-			return ERROR_FILE_IO;
-		}
-
-		(void) snprintf(cmd, sizeof cmd, "%s %s -shared %s -o %s",
-				cc ? cc : "gcc", cflags ? cflags : "",
-				tmp_o, tmp_so);
-
-		break;
-
-	case IMPL_INTERPRET:
-		assert(!"should not reach!");
-		break;
-	}
-
-	if (0 != system(cmd)) {
-		perror(cmd);
-		return ERROR_FILE_IO;
-	}
-
-	if (EOF == fclose(f)) {
-		perror(tmp_c);
-		return ERROR_FILE_IO;
-	}
-
-	if (-1 == unlinkat(-1, tmp_c, 0)) {
-		perror(tmp_c);
-		return ERROR_FILE_IO;
-	}
-
-	if (fd_o != -1) {
-		if (-1 == unlinkat(-1, tmp_o, 0)) {
-			perror(tmp_so);
-			return ERROR_FILE_IO;
-		}
-	}
-
-	h = dlopen(tmp_so, RTLD_NOW);
-	if (h == NULL) {
-		fprintf(stderr, "%s: %s", tmp_so, dlerror());
-		return ERROR_FILE_IO;
-	}
-
-	if (-1 == unlinkat(-1, tmp_so, 0)) {
-		perror(tmp_so);
-		return ERROR_FILE_IO;
-	}
-
-	r->impl = impl;
-
-	switch (impl) {
-	case IMPL_C:
-	case IMPL_VMC:
-		r->u.impl_c.h = h;
-		r->u.impl_c.func = (int (*)(const char *, const char *)) (uintptr_t) dlsym(h, "fsm_main");
-		break;
-
-	case IMPL_VMASM:
-		r->u.impl_asm.h = h;
-		r->u.impl_asm.func = (int (*)(const unsigned char *, size_t)) (uintptr_t) dlsym(h, "fsm_match");
-		break;
-
-	case IMPL_INTERPRET:
-		break;
-	}
-
-	return ERROR_NONE;
-}
-
-enum error_type
-fsm_runner_initialize(struct fsm *fsm, struct fsm_runner *r, enum implementation impl)
-{
-	static const struct fsm_runner zero;
-	struct fsm_dfavm *vm;
-
-	assert(fsm != NULL);
-	assert(r   != NULL);
-
-	/* XXX - minimize or determinize? */
-	if (!fsm_determinise(fsm)) {
-		return ERROR_DETERMINISING;
-	}
-
-	*r = zero;
-
-	switch (impl) {
-	case IMPL_C:
-	case IMPL_VMASM:
-	case IMPL_VMC:
-		return runner_init_compiled(fsm, r, impl);
-
-	case IMPL_INTERPRET:
-		vm = fsm_vm_compile_with_options(fsm, vm_opts);
-		if (vm == NULL) {
-			fsm_free(fsm);
-			return ERROR_COMPILING_BYTECODE;
-		}
-		r->impl = impl;
-		r->u.impl_vm.vm = vm;
-		return ERROR_NONE;
-	}
-
-	return ERROR_INVALID_PARAMETER;
-}
-
-void
-fsm_runner_finalize(struct fsm_runner *r)
-{
-	assert(r != NULL);
-
-	switch (r->impl) {
-	case IMPL_C:
-	case IMPL_VMC:
-		assert(r->u.impl_c.h != NULL);
-		dlclose(r->u.impl_c.h);
-		break;
-
-	case IMPL_VMASM:
-		assert(r->u.impl_asm.h != NULL);
-		dlclose(r->u.impl_c.h);
-		break;
-
-	case IMPL_INTERPRET:
-		assert(r->u.impl_vm.vm != NULL);
-		fsm_vm_free(r->u.impl_vm.vm);
-		break;
-
-	default:
-		assert(!"should not reach");
-	}
-}
-
-int
-fsm_runner_run(const struct fsm_runner *r, const char *s, size_t n)
-{
-	assert(r != NULL);
-	assert(s != NULL);
-
-	switch (r->impl) {
-	case IMPL_C:
-	case IMPL_VMC:
-		assert(r->u.impl_c.func != NULL);
-		return r->u.impl_c.func(s, s+n);
-
-	case IMPL_VMASM:
-		assert(r->u.impl_asm.func != NULL);
-		return r->u.impl_asm.func((const unsigned char *)s, n);
-
-	case IMPL_INTERPRET:
-		assert(r->u.impl_vm.vm != NULL);
-		return fsm_vm_match_buffer(r->u.impl_vm.vm, s, n);
-	}
-
-	assert(!"should not reach");
-	abort();
-}
-
 static enum error_type
 perf_case_run(struct perf_case *c, double *delta)
 {
@@ -793,7 +517,7 @@ perf_case_run(struct perf_case *c, double *delta)
 		}
 	}
 
-	ret = fsm_runner_initialize(fsm, &runner, c->impl);
+	ret = fsm_runner_initialize(fsm, &runner, c->impl, vm_opts);
 	if (ret != ERROR_NONE) {
 		fsm_free(fsm);
 		return ret;
@@ -1013,7 +737,13 @@ main(int argc, char *argv[])
 	opt.consolidate_edges = 1;
 
 	opt.comments          = 1;
-	opt.io                = FSM_IO_GETC;
+
+	/*
+	 * The IO API would depend on c->mt == MATCH_STRING ? FSM_IO_PAIR : FSM_IO_GETC;
+	 * except we read in the entire file below anyway, so we just
+	 * use FSM_IO_PAIR for both. The type of fsm_main depends on the IO API.
+	 */
+	opt.io = FSM_IO_PAIR;
 
 	pause                 = 0;
 	impl                  = IMPL_INTERPRET;
